backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
azure-identity>=1.15.0
azure-search-documents>=11.6.0
openai>=1.12.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic-settings>=2.1.0
